Crate sqlx_postgres

PostgreSQL database driver.

Re-exports

pub use types::PgHasArrayType;

Modules

types

Conversions between Rust and Postgres types.

Structs

PgAdvisoryLock

A mutex-like type utilizing Postgres advisory locks.

PgAdvisoryLockGuard

A wrapper for PgConnection (or a similar type) that represents a held Postgres advisory lock.

PgArgumentBuffer

PgArguments

Implementation of Arguments for PostgreSQL.

PgColumn

PgConnectOptions

Options and flags which can be used to configure a PostgreSQL connection.

PgConnection

A connection to a PostgreSQL database.

PgCopyIn

A connection in streaming COPY FROM STDIN mode.

PgDatabaseError

An error returned from the PostgreSQL database.

PgListener

A stream of asynchronous notifications from Postgres.

PgNotification

An asynchronous notification from Postgres.

PgQueryResult

PgRow

Implementation of Row for PostgreSQL.

PgStatement

PgTransactionManager

Implementation of [TransactionManager] for PostgreSQL.

PgTypeInfo

Type information for a PostgreSQL type.

PgValue

Implementation of Value for PostgreSQL.

PgValueRef

Implementation of ValueRef for PostgreSQL.

Postgres

PostgreSQL database driver.

Enums

PgAdvisoryLockKey

A key type natively used by Postgres advisory locks.

PgErrorPosition

PgSeverity

PgSslMode

Options for controlling the level of protection provided for PostgreSQL SSL connections.

PgTypeKind

PgValueFormat

Traits

PgExecutor

An alias for Executor<'_, Database = Postgres>.

PgPoolCopyExt

Implements methods for directly executing COPY FROM/TO STDOUT on a PgPool.

Type Aliases

PgPool

An alias for Pool, specialized for Postgres.

PgPoolOptions

An alias for PoolOptions, specialized for Postgres.

PgTransaction

An alias for Transaction, specialized for Postgres.