Struct sqlx_postgres::any::Any

pub struct Any;

Opaque database driver. Capable of being used in place of any SQLx database driver. The actual driver used will be selected at runtime, from the connection url.