Trait sqlx_postgres::any::AnyConnectionBackend

pub trait AnyConnectionBackend: Any + Debug + Send + 'static {

    // Required methods
    fn name(&self) -> &str;
    fn close(
        self: Box<Self, Global>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send, Global>>;
    fn ping(
        &mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>;
    fn begin(
        &mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>;
    fn commit(
        &mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>;
    fn rollback(
        &mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>;
    fn start_rollback(&mut self);
    fn shrink_buffers(&mut self);
    fn fetch_many<'q>(
        &'q mut self,
        query: &'q str,
        arguments: Option<AnyArguments<'q>>
    ) -> Pin<Box<dyn Stream<Item = Result<Either<AnyQueryResult, AnyRow>, Error>> + Send + 'q, Global>>;
    fn fetch_optional<'q>(
        &'q mut self,
        query: &'q str,
        arguments: Option<AnyArguments<'q>>
    ) -> Pin<Box<dyn Future<Output = Result<Option<AnyRow>, Error>> + Send + 'q, Global>>;
    fn prepare_with<'c, 'q>(
        &'c mut self,
        sql: &'q str,
        parameters: &[AnyTypeInfo]
    ) -> Pin<Box<dyn Future<Output = Result<AnyStatement<'q>, Error>> + Send + 'c, Global>>
       where 'q: 'c;
    fn describe<'q>(
        &'q mut self,
        sql: &'q str
    ) -> Pin<Box<dyn Future<Output = Result<Describe<Any>, Error>> + Send + 'q, Global>>;

    // Provided methods
    fn cached_statements_size(&self) -> usize { ... }
    fn clear_cached_statements(
        &mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>> { ... }
    fn as_migrate(
        &mut self
    ) -> Result<&mut (dyn Migrate + Send + 'static), Error> { ... }

}

Required Methods

fn name(&self) -> &str

The backend name.

fn close( self: Box<Self, Global> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send, Global>>

Explicitly close this database connection.

This method is not required for safe and consistent operation. However, it is recommended to call it instead of letting a connection drop as the database backend will be faster at cleaning up resources.

fn ping( &mut self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>

Checks if a connection to the database is still valid.

fn begin( &mut self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>

Begin a new transaction or establish a savepoint within the active transaction.

Returns a [Transaction] for controlling and tracking the new transaction.

fn commit( &mut self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>

fn rollback( &mut self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>

fn start_rollback(&mut self)

fn shrink_buffers(&mut self)

Forward to Connection::shrink_buffers().

fn fetch_many<'q>( &'q mut self, query: &'q str, arguments: Option<AnyArguments<'q>> ) -> Pin<Box<dyn Stream<Item = Result<Either<AnyQueryResult, AnyRow>, Error>> + Send + 'q, Global>>

fn fetch_optional<'q>( &'q mut self, query: &'q str, arguments: Option<AnyArguments<'q>> ) -> Pin<Box<dyn Future<Output = Result<Option<AnyRow>, Error>> + Send + 'q, Global>>

fn prepare_with<'c, 'q>( &'c mut self, sql: &'q str, parameters: &[AnyTypeInfo] ) -> Pin<Box<dyn Future<Output = Result<AnyStatement<'q>, Error>> + Send + 'c, Global>>where 'q: 'c,

fn describe<'q>( &'q mut self, sql: &'q str ) -> Pin<Box<dyn Future<Output = Result<Describe<Any>, Error>> + Send + 'q, Global>>

Provided Methods

fn cached_statements_size(&self) -> usize

The number of statements currently cached in the connection.

fn clear_cached_statements( &mut self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + '_, Global>>

Removes all statements from the cache, closing them on the server if needed.

fn as_migrate(&mut self) -> Result<&mut (dyn Migrate + Send + 'static), Error>