sqlx_postgres

Struct PgAdvisoryLockGuard

pub struct PgAdvisoryLockGuard<'lock, C: AsMut<PgConnection>> { /* private fields */ }

A wrapper for PgConnection (or a similar type) that represents a held Postgres advisory lock.

Can be acquired by PgAdvisoryLock::acquire() or PgAdvisoryLock::try_acquire(). Released on-drop or via Self::release_now().

Note: Release-on-drop is not immediate!

On drop, this guard queues a pg_advisory_unlock() call on the connection which will be flushed to the server the next time it is used, or when it is returned to a PgPool in the case of PoolConnection<Postgres>.

This means the lock is not actually released as soon as the guard is dropped. To ensure the lock is eagerly released, you can call .release_now().await.

Implementations

impl<'lock, C: AsMut<PgConnection>> PgAdvisoryLockGuard<'lock, C>

pub async fn release_now(self) -> Result<C>

Immediately release the held advisory lock instead of when the connection is next used.

An error should only be returned if there is something wrong with the connection, in which case the lock will be automatically released by the connection closing anyway.

If pg_advisory_unlock() returns false, a warning will be logged, both by SQLx as well as the Postgres server. This would only happen if the lock was released without using this guard, or the connection was swapped using std::mem::replace().

pub fn leak(self) -> C

Cancel the release of the advisory lock, keeping it held until the connection is closed.

To manually release the lock later, see PgAdvisoryLock::force_release().