Struct sqlx_postgres::types::PgMoney

pub struct PgMoney(pub i64);

The PostgreSQL MONEY type stores a currency amount with a fixed fractional precision. The fractional precision is determined by the database’s lc_monetary setting.

Data is read and written as 64-bit signed integers, and conversion into a decimal should be done using the right precision.

Reading MONEY value in text format is not supported and will cause an error.

locale_frac_digits

This parameter corresponds to the number of digits after the decimal separator.

This value must match what Postgres is expecting for the locale set in the database or else the decimal value you see on the client side will not match the money value on the server side.

For most locales, this value is 2.

If you’re not sure what locale your database is set to or how many decimal digits it specifies, you can execute SHOW lc_monetary; to get the locale name, and then look it up in this list (you can ignore the .utf8 prefix): https://lh.2xlibre.net/values/frac_digits/

If that link is dead and you’re on a POSIX-compliant system (Unix, FreeBSD) you can also execute:

$ LC_MONETARY=<value returned by `SHOW lc_monetary`> locale -k frac_digits

And the value you want is N in frac_digits=N. If you have shell access to the database server you should execute it there as available locales may differ between machines.

Note that if frac_digits for the locale is outside the range [0, 10], Postgres assumes it’s a sentinel value and defaults to 2: https://github.com/postgres/postgres/blob/master/src/backend/utils/adt/cash.c#L114-L123

Tuple Fields
0: i64

The raw integer value sent over the wire; for locales with frac_digits=2 (i.e. most of them), this will be the value in whole cents.

E.g. for select '$123.45'::money with a locale of en_US (frac_digits=2), this will be 12345.

If the currency of your locale does not have fractional units, e.g. Yen, then this will just be the units of the currency.

See the type-level docs for an explanation of locale_frac_units.