Module sqlx_postgres::any

Modules

driver

types

Conversions between Rust and standard SQL types.

Structs

Any

Opaque database driver. Capable of being used in place of any SQLx database driver. The actual driver used will be selected at runtime, from the connection url.

AnyArgumentBuffer

AnyArguments

AnyColumn

AnyConnectOptions

Opaque options for connecting to a database. These may only be constructed by parsing from a connection url.

AnyConnection

A connection to any SQLx database.

AnyQueryResult

AnyRow

AnyStatement

AnyTransactionManager

AnyTypeInfo

AnyValue

AnyValueRef

Enums

AnyKind

AnyTypeInfoKind

AnyValueKind

Constants

DRIVER

Traits

AnyConnectionBackend

AnyExecutor

An alias for Executor<'_, Database = Any>.

Type Definitions

AnyPool

AnyPoolOptions